Ch14FastApiEmp/requirements.txt
fastapi~=0.115.12
uvicorn[standard]
asyncpg~=0.30.0
python-dotenv~=1.1.0

pydantic~=2.11.4